requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.1
